+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/MatrixMessengerWorker.kt
@@ -1,0 +1,5 @@
+package de.connect2x.trixnity.messenger
+
+fun interface MatrixMessengerWorker {
+    suspend operator fun invoke()
+}