+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/multi/MatrixMessengerWorker.kt
@@ -1,0 +1,5 @@
+package de.connect2x.trixnity.messenger.multi
+
+fun interface MatrixMultiMessengerWorker {
+    suspend operator fun invoke()
+}